+++ types/markdown-it-container/markdown-it-container-tests.ts
@@ -1,11 +1,13 @@
 import * as MarkdownIt from "markdown-it";
+import * as Token from "markdown-it/lib/token";
+
 import MarkdownItContainer = require("markdown-it-container");
 
 const md = new MarkdownIt();
 
 md.use(MarkdownItContainer, 'spoiler', {
 	validate: (params: any) => params.trim().match(/^spoiler\s+(.*)$/),
-	render: (tokens: MarkdownIt.Token[], index: number) => {
+	render: (tokens: Token[], index: number) => {
 		const match = tokens[index].info.trim().match(/^spoiler\s+(.*)$/);
 		const onClick =
 			"this.parentNode.classList.toggle('_expanded');" +
