+++ types/markdown-it-lazy-headers/index.d.ts
@@ -4,7 +4,7 @@
 // Definitions: https://github.com/DefinitelyTyped/DefinitelyTyped
 // TypeScript Version: 2.3
 
-import { MarkdownIt } from 'markdown-it';
+import * as MarkdownIt from 'markdown-it';
 
 declare function lazyheaders(md: MarkdownIt): void;
