+++ types/markdown-it/lib/parser_core.d.ts
@@ -1,0 +1,12 @@
+import * as MarkdownIt from ".";
+import * as Ruler from "./ruler";
+import * as Token from "./token";
+
+export = ParserCore;
+
+declare class ParserCore {
+    process(state: any): void;
+    ruler: Ruler;
+}
+
+declare module ParserCore {}
